perf(hotel): change folder
