--- apps/2CD/hotel/frontend/src/app/homeTest/page.tsx
@@ -1,10 +1,0 @@
-import HotelCard from './_features/hotel-card';
-
-const HomePageTest = () => {
-  return (
-    <>
-      <HotelCard />
-    </>
-  );
-};
-export default HomePageTest;
