+++ apps/2CD/hotel/frontend/src/app/page.tsx
@@ -2,24 +2,17 @@
 /* eslint-disable @next/next/no-img-element */
 /* eslint-disable complexity */
 'use client';
-import Header from "@/components/header/Header";
-import FooterRegular from "@/components/footer/footer-regular";
-import FooterReserve from "@/components/footer/footer-reserve";
-import FooterCheckIn from "@/components/footer/footer-check-in";
+import Header from '@/components/header/Header';
+import FooterReserve from '@/components/footer/footer-reserve';
+import HotelCard from '@/components/hotel/hotel-card';
 const Page = () => {
-
   return (
     <div className="h-screen w-full ">
       <Header></Header>
-      <main className="h-full">
-      </main>
+      <HotelCard />
       <FooterReserve></FooterReserve>
-      <FooterCheckIn></FooterCheckIn>
-      <FooterRegular></FooterRegular>
     </div>
   );
 };
 
 export default Page;
-
-
