+++ apps/2CD/hotel/frontend/src/components/hotel/amenities-section.tsx
