+++ apps/2CD/hotel/frontend/src/components/hotel/most-booked-hotel.tsx
@@ -1,0 +1,11 @@
+/* eslint-disable unicorn/filename-case */
+'use client ';
+
+export const MostBookedHotel = () => {
+  return (
+    <>
+      <div></div>
+      <div></div>
+    </>
+  );
+};
